% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMboostE.R
\name{JMboost}
\alias{JMboost}
\title{Main function to carry out boosting for a joint model!}
\usage{
JMboost(y, Xl = NULL, Xs = NULL, Xls = NULL, T_long, T_surv, delta,
  id, alpha = 1, lambda = 1, time.effect = TRUE, mstop_l, mstop_s,
  mstop_ls, nyl = 0.1, nys = 0.3, nyls = 0.1, nyr = 0.1)
}
\arguments{
\item{y}{vector containing the longitudinal outcome}

\item{Xl}{design matrix for the longitudinal part containing time-varying covariates}

\item{Xs}{design matrix for the survival part containing one measurement per individual and covariate}

\item{Xls}{design matrix for the shared part containing time-independent covariates (duplicated values per individual)}

\item{T_long}{longitudinal time points}

\item{T_surv}{observed survival times}

\item{delta}{censoring indicator}

\item{id}{id vector labeling the longitudinal outcome/design matrices with corresponding individuals}

\item{alpha}{starting value for the association parameter}

\item{lambda}{starting value for the baseline hazard}

\item{time.effect}{logical, whether a shared time.effect shall be included}

\item{int}{starting value for the intercept in the longitudinal submodel}

\item{mstop_i}{number of boosting iterations per boosting step, i = l, s, ls}

\item{nyi}{step length per boosting step, i = l, s, ls}
}
\description{
Main function to carry out boosting for a joint model!
}
