% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simJME.R
\name{simJM}
\alias{simJM}
\title{Function to simulate a data structure suitable for joint modeling!}
\usage{
simJM(n, n_i, betal = 0, betas = 0, betat = 0, betals = 0, int,
  alpha, lambda, sigma2, noninfl = 0, noninfs = 0, noninfls = 0,
  high.dim = FALSE)
}
\arguments{
\item{n}{number of individuals}

\item{n_i}{initial number of longitudinal measurements per individual}

\item{betal}{coefficient vector for the longitudinal model}

\item{betas}{coefficient vector for the survival model}

\item{betat}{value for the time effect}

\item{betals}{coefficient vector for the shared model}

\item{int}{value for the intercept}

\item{alpha}{value for the association parameter}

\item{lambda}{value for the baseline hazard}

\item{sigma2}{value for the model error}

\item{noninfi}{number of additional, non-informative covariates per predictor, i = l, s, ls}
}
\description{
Function to simulate a data structure suitable for joint modeling!
}
